requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.7.5
boto3==1.20.10
botocore==1.23.10
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
cytoolz==0.11.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
fonttools==4.28.2
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jellyfish==0.8.9
jieba3k==0.35.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
langcodes==3.3.0
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
newspaper3k==0.2.8
nltk==3.6.5
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.10.0
Pympler==0.9
pyparsing==3.0.6
pyphen==0.11.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
response==0.5.0
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
streamlit==1.2.0
terminado==0.12.1
testpath==0.5.0
textacy==0.11.0
thinc==8.0.13
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
